+++ src/prog2/model/CamiAsfalt.java
@@ -2,7 +2,7 @@ package prog2.model;
 
 public class CamiAsfalt extends AccesAsfalt{
 
-    public CamiAsfalt(String nom, boolean estat, long areaAsfalt){
+    public CamiAsfalt(String nom, boolean estat, float areaAsfalt){
         super(nom,estat,areaAsfalt);
         this.accessibilitat = false;
     }
