+++ src/prog2/model/CamiTerra.java
@@ -2,7 +2,7 @@ package prog2.model;
 
 public class CamiTerra extends AccesTerra{
 
-    public CamiTerra(String nom, boolean estat, long longitud){
+    public CamiTerra(String nom, boolean estat, float longitud){
         super(nom,estat,longitud);
         this.accessibilitat = false;
     }
